Implement OCI registry basic auth
This changeset adds a `--docker-config-json-path` that will be used
internally to match registry hostnames. The file has to be of the
form:

```json
{
  "auths": {
    "registry1.company.com": {
      "auth": "<base64 of username:password>"
    },
    "local-registry:5000": {
      "auth": "<base64 of other-username:other-password>"
    }
  }
}
```

Whenever an OCI artifact will be pulled from a host that matches a key
on the `auths` map, the policy server will try to authenticate against
the registry with the given details.