Allow user to choose where to download policies
Prior to this commit the policies would be downloaded into the local execution
directory. This commit allows the user to specify where the policies
files downloaded from a HTTP(s) server/OCI registry should be saved.

This is needed to run policy-server containerized. Without this patch
the server would try to download the policies to `/usr/local/bin`, which
is a bad thing; plus it's not writable by the unprivileged user running
inside of the container.

Signed-off-by: Flavio Castelli <fcastelli@suse.com>