Implement cluster context awareness
This changeset allows the policy server to retrieve a list of well
known resources (namespaces, ingresses and services) and keeps them
updated by listing every five seconds.

This changeset also implements a callback on the policy evaluator that
allows the SDK's to consume this list of resources as a regular
string, that they can unmarshal to provide to the native language the
specific SDK targets.

By implementing the cluster context awareness inside the policy
evaluator, the listing of resources that is done in a dedicated
thread, expands the `policy-server` and the `policy-testdrive` as
well. This allows the `policy-testdrive` to automatically fetch this
list of well-known resources when it starts, so it can also be used to
perform complex tests that require polling information from a
Kubernetes cluster.

Some things that can be improved:

* Generalize types instead of relaying on a list of well known type of
  API resources. By generalizing we allow to also consume custom
  resources or specific Kubernetes types that are important to
  certain policies.

* Do not fetch & poll any kind of resource; delegate first fetch until
  first request is made by a policy, then include this type of
  resource on a watch.

* As specified on the previous point, use watchers instead of active
  polling.