Initial `kwctl` implementation
This initial implementation allows the user to run policies similar to
what `docker run` allows to do with containers.

To be squashed: make pull and run top level commands

Initial implementation of kwctl

This initial version implements the following commands:

- `pull`: it can pull policies from OCI registries or HTTP servers. By
  default it will place the WebAssembly module in `~/.kubewarden`.

- `run`: performs a `pull` beforehand, and then creates a WebAssembly
  runtime, initiates the module with the given `--settings`, and then
  passes the `--request` to it, evaluating the provided request.

- `policies`: lists the downloaded policies.

The storage implementation is filesystem based, and does not contain
any kind of index or embedded database. The cache filesystem structure
is as follows:

```
<root> (~/.kubewarden by default)
  - <schema>
    - <host>
      - <path>
        - <path>
          - <module>:<tag>.wasm
```

For a `<schema>://<host>/<path>/<path>/<module>:<tag>` URI.