massive refactoring of the main init
Change the way the worker pool is initialized.
Now the worker pool performs the bootstrap once it receives a message sent
over a tokio channel.

This allows to organize the code inside of the `main` function in two
distinctive areas:

  1. Setup the synchronous code: start the worker pool thread, but do
  not bootstrap the workers.
  2. Setup the tokio runtime
  3. Have a single block where all the code relying on a tokio runtime (
    distributed tracing initialization, policy fetching, web server) is
    executed

This distinct separation was needed to properly init some tracing
collectors like Jaeger and OpenTelemetry, which needed a tokio Runtime
to be available.

As a nice side effect, the whole code is much more clear and well
organized.