Default mutating webhook prevent resource to be deleted.
Default mutating webhook prevents resources to be deleted.

When deleting a policy server the policies are deleted. In this process,
the webhook register to mutate the `clusteradmissionpolicies` receive a
request. The function used to handle this request is the default one
which adds a `kubewarden` finalizer in the resource preventing the
policy to be removed and consequently the policy server is not deleted
as well. To work around this the default webhook code checks if the
resource has a deletion timestamp which indicates that the resource
should be deleted and does not add the default finalizer.