Expose the get OCI manifest digest callback
Allow the Wasm guests to obtain the manifest's digest of
an OCI object (be it a container or an artifact).

Fetching information from a OCI registry requires the execution
of asynchronous code. Unfortunately the `host_callback` function
is a synchronous function.

Creating a dedicated tokio runtime (even by reusing the "current thread"
of the `host_callback` function) is not going to work. This would cause
tokio to cause a `panic` at runtime.

The solution of this problem consists of using a tokio channel that can
send the evaluation request from the synchronous world (the `host_callback`
function) to an `asynchronous` "worker" (something created with a tokio
`spawn` for example).

The response from the asynchronous world is then sent back to the
synchronous one by using a dedicated tokio `oneshot` channel.

Changes
=======

On top of exposing the new waPC host function, and adding all the tokio
channels. The following changes have been done:

  * The creation of a `PolicyEvaluator` is now done by using a dedicated
    builder: `PolicyEvaluatorBuilder`.
  * The code executing `PolicyEvaluator` (kwctl, policy-server) must
    prepare the worker that can handle the evaluation requests coming
    from the waPC guests.

Signed-off-by: Flavio Castelli <fcastelli@suse.com>