refactor: Use Option<&T> for kwctl functions
Use Option<&T> instead of &Option<T> or Option<T> in kwctl functions.

This is not exhaustive as policy-fetcher still has pub functions that
expect owner objects instead of references. With this commit we use
references on the kwctl code, and use cloned objects just on the calls
to policy-fetcher functions.