Registry push: return immutable reference
Change the API of `Registry::push` to return a value inside of its
`Result` response.

The method now returns the immutable reference of the policy that has
been pushed.

This information is important, because it allows kwctl users to use
cosign (from Sigstore) to sign the **actual** policy that has just been
pushed to the repo.

Signing policies by tag is not secure, there could be race conditions or
attacks that could lead the user to sign a different content. Signing
by immutable ref is not affected by these issues.

Signed-off-by: Flavio Castelli <fcastelli@suse.com>