Do not flatten settings
Do not flatten the `settings` field that would behave as a "catch-all"
for unknown attributes.

This would silently make the `policy-server` forward certain
attributes to policies when keys change on the configuration. We
should plan for such changes and have a migration path.

Policies should only receive as `settings` what is intended as such.