feat: implement policy modes
Policies can be configured to either be in `monitor` or `protect`
modes. When a policy is in `protect` mode, it will accept, reject or
mutate requests.

When a policy is in `monitor` mode, it will always accept all
requests, but will log what would have been its decision if it had
been in `protect` mode. Also, in this mode, mutations will be erased
on the final answer and logged.

This is helpful if you want to check how a policy would behave before
it starts to take decisions on a Kubernetes cluster.