Implement monitor mode
Add a new attribute to the `ClusterAdmissionPolicy` resource called
`mode`, that can accept values `protect` or `monitor`.

It is allowed to transition a policy from a `monitor` mode to a
`protect` mode, but it's not allowed to transition from `protect` to
`monitor`. This is so by giving UPDATE RBAC permissions to a user we
know that they can transition the policy only to a more strict runtime
configuration.

If it's desired to transition a policy from `protect` to `monitor`, it
is required to have DELETE RBAC permissions on the
`ClusterAdmissionPolicy` resource, so the policy can be deleted and
recreated again in `monitor` mode.