Does not update the old object of an update event.
Kubewarden controller has a problem when the user delete a Cluster Admission
Policy. The controller tries to update the status of the old object of the
update event which adds the delete timestamp in the resource. This fails
because the old object is not the current version of the resource, causing a
conflict and a ugly stacktrace in the logs.

To solve this, this commit rewrite the functions used to watch such events and
changes the update event handler to change status only in the new object. As
well as moves the delete handler to a separate function.