Remove wasmtime dependency
The wasmtime dependency was introduced to have a reliable way to detect
corrupted Wasm modules.
That was done because GitHub Container Registry sometimes returns an
html page instead of a Wasm module.

The wasmtime dependency is a bit of an overkill, plus it makes
maintenance a bit harder because it could cause dependency resolution
conflicts between policy-fetcher and other kubewarden crates that
**need** wasmtime.

This commit removes the wasmtime dependency and instead identifies valid
Wasm files by looking at their "magic bytes".

Signed-off-by: Flavio Castelli <fcastelli@suse.com>
Co-authored-by: Rafael Fernández López <ereslibre@kde.org>