feat: Pretty print unsatisfied `config::Signature`
Refactor `verify_signatures_against_config()` from calling
`cosign::verify_constraints()` with a vector of constraints, to calling it
with only 1, and move the iteration outside. This essentially
re-implements `cosign::verify_constraints()` in
`verify_signatures_against_config()`, and allows us to create a vector
of unsatisfied config::Signatures for printing.

This will allow us in the future to paralelize the signatures_all_of
iter, too.

Signed-off-by: Víctor Cuadrado Juan <vcuadradojuan@suse.de>
Co-Authored-by: Flavio Castelli <fcastelli@suse.com>
Signed-off-by: Flavio Castelli <fcastelli@suse.com>