Tests checks if policies are active.
After some recent changes in the controller, the policies status set to
"active" before the policy is consider as unique reachable. Which causes
the policy to pass over the "pending" status faster. Therefore, the test
need to be updated to check if the policy is "active" instead of
"pending". Otherwise, the tests will fail.