Remove resource name from metric.
Removes the resource_name field from the metrics. The Prometheus
documentation recommends avoid adding fields with high cardinality as
metric label. This is necessary because every single combination in the
labels values generate another time series. Which can increase the
amount of data stored.

Besides that, removing the resource_name from the metric make the
queries used to monitor the environment simpler and easy to visualize.