deps: Set syntect to only use `parsing` feature
We don't need the rest of the features as our use of syntect is minimal:
`syntect::parsing::SyntaxSet::load_defaults_newlines()` for mdcat.

This reduces deps and compilation time, plus removes the dependency on
`plist`, which on its own depends on `xml-rs`.
`xml-rs` is unmaintained, a RUSTSEC advisory has been issued for it:
https://rustsec.org/advisories/RUSTSEC-2022-0048.html