fix: have a working container image
Ensure the resulting container image is working. Since we introduced the
SBOM generation we caused a regression that lead the `policy-server`
binary to be added to the final image under
`/policy-server/policy-server-<arch>`.

That caused the resulting image to not work, because the entrypoint was
a directory instead of a file.

This commit fixes the issue

Signed-off-by: Flavio Castelli <fcastelli@suse.com>