feat: Add callback `crypto/v1/is_certificate_trusted`
Given a cert and a cert chain, verifies the trustworthiness of the cert
against the passed cert chain, and the validity of the cert (if it is
expired, and if it is not being used before it's validity time).
If no chain is passed, only the cert validity against the current time
is checked.

Add new module `callback_handler::crypto`, that implements the cert
verification.

Depend on picky crate, same version as sigstore-rs.

Signed-off-by: Víctor Cuadrado Juan <vcuadradojuan@suse.de>