fix: remove confusing messages from bench output
The bench command should not attempt to write the benchmarks results to
disk by default.
The crate we're using has a hard-coded path which exists only when
inside of a Rust project.

Prior to this commit, the following warning messages (shown in red) were
printed when the path was not found:

```console
Failed to read last sample, cause Could not find target directory to place output
Failed to write sampling data, cause: Could not find target directory to place output
```

This can mislead/scare end users.

This commit disables this behavior and introduces a new flag that can be
used to write the results back to disk.

Signed-off-by: Flavio Castelli <fcastelli@suse.com>