wapc policy: reset evaluator after epoch interruption
TL;DR: after code execution is interrupted because of an
epoch deadline being reached, we have to reset the waPC host
to ensure further invocations of the policy work as expected.

The waPC host is using the wasmtime_provider, which internally
uses a wasmtime::Engine and a wasmtime::Store.
The Store keeps track of the stateful data of the policy. When an
epoch deadline is reached, wasmtime::Engine stops the execution of
the wasm guest. There's NO CLEANUP code called inside of the guest.
It's like unplugging the power cord from a turned on computer.

When the guest function is invoked again, the previous state stored
inside of wasmtime::Store is used.
That can lead to unexpected issues. For example, if the guest makes
uses of a Mutex, something like that can happen (I've witnessed that):

* Guest code 1st run:
  - Mutex.lock
* Host: interrupt code execution because of epoch deadline
* Guest code 2nd run:
  - The Mutex is still locked, because that's what is stored inside
    of the wasmtime::Store
  - Guest attempts to `lock` the Mutex -> error is raised

The guest code will stay in this broken state forever. The only
solution to that is to reinitialize the wasmtime::Store.
It's hard to provide a facility for that inside of WapcHost, because
epoch deadline is a feature provided only by the wasmtime backend.
Hence it's easier to just recreate the wapc_host associated with this
policy evaluator

On top of this new code, a small refactor of the `runtimes/mod.rs`
was done to splits wapc and burrego code into dedicated `.rs` files.

Signed-off-by: Flavio Castelli <fcastelli@suse.com>