fix: policy-server pull from authenticated registries
After the 1.5.0 release, Policy Server changed the source code that
parses the docker credentials.
The new code comes from a 3rd party Rust crate: docker_credentials.

The library looks for for a file called`docker.config` file under the
`~/.docker` directory.

The Kubernetes secret holding the docker credentials was mounted to
`/home/kubewarden/docker/.dockerconfigjson`, causing Policy Server to
not be able to find the credentials. Hence, pulling from a registry
using authorization resulted in an anonymous pull, which then lead to
a pull error.

This code changes the Deployment of Policy Server to mount the secret
under the expected location: `/home/kubewarden/.docker/config.json`.

Signed-off-by: Flavio Castelli <fcastelli@suse.com>