fix: allow kubectl debug to be used against Policy Server
After the 1.5.0 release, some changes have been made to the security
context of the Pod running Policy Server. By default the Pod had a
securityContext that prevented the creation of containers whose
entrypoint was executed by the root user.

This restriction makes impossible to use kubectl debug against the
policy-server container. That's because the sidecar debug container
requires to run as root in oder to perform debug operations.

This commit removes the default securityContext added to the Policy
Server Pod. We leave the ability to the end user to specify his own
securityContext to be applied to the Policy Server Pod.

Note well: this change does not make the Policy Server deployment less secure.
As a matter of fact, the Policy Server Pod is made by 1 single
container: the one running the actual Policy Server. This container has
its own securityContext which already enforces the root user to NOT be
used.

The securityContext we previously added to the Pod was affecting the
sidecars being added to the Pod. Potentially breaking them.

Signed-off-by: Flavio Castelli <fcastelli@suse.com>