fix: controller failed to update policy server service.
The controller is not able to update the policy server service because
it always call the create function in the reconciliation loop.
Therefore, if the user edits the policy server enabling telemetry, the
service will not be updated adding the metric port in the service.

To fix that, this change adds a call to the update function when the
creation failed due an "already exists" error.

Signed-off-by: José Guilherme Vanz <jguilhermevanz@suse.com>