refactor: change how the container image is built
Prior to this commit, building the container image was done performing a
complex cross-compilation build of the policy-server binary.
That required a special environment to be setup inside of the GitHub
action. This environment was a bit scary to maintain, it required
a specific set of repositories and workarounds in place, all of them
done inside of our GitHub action.

Because of that, performing a local build of the container image was not
doable.

This change allows to build the container image doing a simple "docker
build". Moreover, the complex build environment is no longer needed.

Signed-off-by: Flavio Castelli <fcastelli@suse.com>