fix: release linux binaries
The build process of linux binaries failed during the release process.
That happened because one of the ubuntu repositories added by the build
steps is no longer reachable.

This commit changes the GH action to perform the cross building using
the rust-cross project. This hugely simplifies the action, making
everything more maintainable

Signed-off-by: Flavio Castelli <fcastelli@suse.com>