fix: improve error message when AdmissionRequest is not valid
Due to recent changes, the AdmissionRequest validation is stricter. Some
of our e2e test fixtures are missing required values, making the test
fail (which is the right thing to do).

This commit improves the error message given when the AdmissionRequest
object is failing the validation. This makes everything less cryptic.

Signed-off-by: Flavio Castelli <fcastelli@suse.com>