fix: policy server resource deleted before policies
There is an issue in the policy server controller which is causing
inconsistencies when the policy server is deleted. It's possible, due
the concurrent nature of events and controllers loops, where the
policies ended in a inconsistent state. Where the policy and its
webhooks are register in the cluster. But the policy server which should
run it is gone. Causing an issue when the control plane tries to contact
the policy server.

This is caused because when the policy server controller deletes all the
policy server policies, it filters out policies which have the deletion
timestamp field. Once no policies are found, in other words, when no
policies without deletion timestamp are found, the policy server
resources are deleted. However, the deletion timestamp does not ensure
that the resource is not register in the cluster. It tells that the
resource is marked to be deleted by the control plane. Which means that
we can reach a state where the policy server is deleted before the
policies. Causing issue when the control plane tries to send request to
the webhooks.This commit fixes that by removing the filter used to get
the policies in the deletion reconciliation loop.

Signed-off-by: José Guilherme Vanz <jguilhermevanz@suse.com>