feat: use owner reference with policy server resources.
Updates the controller to set owner references for all the resources
created to make the policy server works. Which includes: service,
configmap, deployment and secret. Therefore, we can remove the
reconciliation loop code to remove this resources when the user deletes
the policy server and rely on Kubernetes garbage collector to do the
job.

Signed-off-by: José Guilherme Vanz <jguilhermevanz@suse.com>