fix: patch webhooks not update them.
In a recent change on how we create or update a webhooks the controller
started to use the helper function `CreateOrUpdate` to create or update
the webhooks. However, the RBAC definition used by the controller allow
only to patch the webhooks. Therefore, the controller failed when it
trys to update the resources. This commit fixes the issue by using the
`CreateOrPatch` helper function instead. The behaviour is the same but
it uses the patch verb to "update" the webhooks resources.

Signed-off-by: José Guilherme Vanz <jguilhermevanz@suse.com>