fix: do not check if affinity is empty.
The policy server reconciler check if the policy server spec has a non
empty affinity. If that's the case, it set the affinity in the policy
server deployment. This allow users to set the affinity manually in the
deployment. This should not be allowed. The policy server CRDS should be
the interface to configure the affinity. Therefore, that check is
removed and the affinity from the policy server spec is always set in
the deployment.

Signed-off-by: José Guilherme Vanz <jguilhermevanz@suse.com>