fix: make cert rotation detection more reliable
When detecting certificate and key changes, rely on the proper inotify
message.

We have to wait for the `CLOSE_WRITE` event to happen to be sure all the
contents of the cert/key have been written and flushed to disk.

Otherwise there's the risk of loading some incomplete data, leading to a
runtime panic.

This was detected thanks to a flaky test.

Signed-off-by: Flavio Castelli <fcastelli@suse.com>