chore(ci): allow slsactl verification.
Updates the release.yml file to allow users verify the attestation
content of the policy server using slsactl command. slsactl does not
support the verification of the artefacts when it's build using sub
workflow files. This is fixed by copying the content of the
container-build.yml file into the release.yml file.

Signed-off-by: José Guilherme Vanz <jguilhermevanz@suse.com>