fix: add more logging
Add more logging statements when pushing to an insecure registry.

We had a user that was struggling to understand why pushing to a certain
registry didn't work despite `insecure` being turned on.

We finally figured out that was happening because of a reverse proxy was
misconfigured. However, the error obtained from the insecure proxy was
silently swallowed.

Signed-off-by: Flavio Castelli <fcastelli@suse.com>