dev: fix local tilt development
Do not configure the registry that holds the development image of
kubewarden-controller as the place from which all the images of the
`kubewarden-controller` helm chart have to be pulled from.
This would cause the deployment to fail, since all the other images
(like kubectl, audit-scanner) are not mirrored on the local registry.

Instead, patch the image attribute of the Deployment.

Signed-off-by: Flavio Castelli <fcastelli@suse.com>