refactor(policy-server): Keep old policy-server label key
Revert the change that set a valid domain for the PolicyServerLabelKey,
instead of just `kubewarden`, as this would make upgrades break without
further reconciler changes.

This label is used to tie Pods, ConfigMaps, Services with a specific
PolicyServer by the controller, and will be removed in future policy
lifecycle refactoring.

Signed-off-by: Víctor Cuadrado Juan <vcuadradojuan@suse.de>